move localization code to api module
