+++ module/crypto-demo/src/main/java/org/example/age/module/crypto/demo/DemoSiteVerifiedUserLocalizer.java
@@ -5,7 +5,7 @@ import jakarta.inject.Singleton;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import org.example.age.api.VerifiedUser;
-import org.example.age.service.module.crypto.Localization;
+import org.example.age.api.crypto.Localization;
 import org.example.age.service.module.crypto.SiteVerifiedUserLocalizer;
 
 /**
