+++ service/src/test/java/org/example/age/service/testing/crypto/FakeAvsVerifiedUserLocalizer.java
@@ -7,9 +7,9 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import org.example.age.api.VerifiedUser;
+import org.example.age.api.crypto.Localization;
 import org.example.age.api.crypto.SecureId;
 import org.example.age.service.module.crypto.AvsVerifiedUserLocalizer;
-import org.example.age.service.module.crypto.Localization;
 
 /** Fake implementation of {@link AvsVerifiedUserLocalizer}. */
 @Singleton
