+++ service/src/test/java/org/example/age/service/testing/crypto/FakeSiteVerifiedUserLocalizer.java
@@ -5,8 +5,8 @@ import jakarta.inject.Singleton;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import org.example.age.api.VerifiedUser;
+import org.example.age.api.crypto.Localization;
 import org.example.age.api.crypto.SecureId;
-import org.example.age.service.module.crypto.Localization;
 import org.example.age.service.module.crypto.SiteVerifiedUserLocalizer;
 
 /** Fake implementation of {@link SiteVerifiedUserLocalizer}. */
